fixed e2e tests and class renamed
